= 3.0.7 = * Fixed content placement and product store * Added Recent Searches Widget * Option to add categories to No Resaults on product searches, and Top Products widget (will help with niche sites) * Added option to change grid view image sizes * Fixed some theme issues * Plugin will now try to create prosperent_cache and prosperent-themes directories * Added button to clear cache under General Options (only visible when Caching is enabled) * Cache times are now variables for easy changing in the prosperent-suite.php file * Recent Searches show up at the bottom of Product Search settings, to allow you to delete unwanted searches * Fixed bug with '/' in brand/merchant names * Fixed bug that some servers weren't rendering short code .phtml * Fixed WordPress Demo URLs * Updated README.txt * [Prosperent Demo](http://wordpress.prosperentdemo.com/) has been redone to offer a more in depth guide
git-svn-id: https://plugins.svn.wordpress.org/prosperent-suite/trunk@881976 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/views/prosperadmin/themes-phtml.php
@@ -2,11 +2,11 @@
 require_once(PROSPER_MODEL . '/Admin.php');
 $prosperAdmin = new Model_Admin();
 
-$prosperAdmin->adminHeader( __( 'Themes and CSS', 'prosperent-suite' ), true, 'prosperent_themescss_options', 'prosper_themes' );
+$prosperAdmin->adminHeader( __( 'Themes', 'prosperent-suite' ), true, 'prosperent_themescss_options', 'prosper_themes' );
+
+echo '<p class="prosper_settingDesc">' . __( 'The main idea behind themes is to allow people to edit their stores how they see fit. Whether they want to change the CSS or alter the entire store front. Themes will make that possible, even while the plugin gets updated.<br><br>Go to <a href="http://wordpress.prosperentdemo.com/themes/">WordPress Prosperent Demo: Themes</a> for more information and to download an Example Theme.', 'prosperent-suite' ) . '</p>';
 
 echo '<h2 class="prosper_h2">' . __( 'Change the Theme for the Search', 'prosperent-suite' ) . '</h2>';
-echo __( '<ol><li>Create a <strong>prosperent-themes</strong> directory inside your <strong>wp-content</strong> directory.</li><li>Import new theme directories.<a href="#" class="prosper_tooltip"><img border="0" src="' . PROSPER_IMG . '/help.png"><span>New theme directories should contain one or all of the following: <br>&bull;&nbsp;<strong>product.php</strong><br>&bull;&nbsp;<strong>productPage.php</strong><br>&bull;&nbsp;a <strong>css file</strong> <br>(The CSS file can be named anything, the plugin is looking for the extension .css).</span></a></li><li>Select the desired theme below.</li></ol>', 'prosperent-suite' );
-echo '<p class="prosper_desc">' . __( "", 'prosperent-suite' ) . '</p>';
 
 if (file_exists(PROSPER_THEME))
 {
@@ -20,18 +20,31 @@ if (file_exists(PROSPER_THEME))
 	foreach ($themesDir as $theme)
 	{
 		$themes[$theme] = ucwords($theme);
-		
 	}
 
 	echo $prosperAdmin->select( 'Set_Theme', __( 'Select Theme to Use', 'prosperent-suite' ),  $themes);
-	echo '<p class="prosper_descb">' . __( "Select <strong>Default</strong> if you want to use the default theme.", 'prosperent-suite' ) . '</p>';
+	echo '<p class="prosper_desc">' . __( "Select <strong>Default</strong> if you want to use the default theme.", 'prosperent-suite' ) . '</p>';
 }
 else
 {
+	shell_exec('mkdir ' . PROSPER_THEME);
+	
+	if (!file_exists(PROSPER_THEME))
+	{
 		echo '<div class="update-nag" style="padding:6px 0;">';
-		echo _e( '<span style="font-size:14px; padding-left:10px;">There is an issue with your <strong>prosperent-themes</strong> directory. </span><br>', 'my-text-domain' );
-		echo _e('<span style="font-size:14px; padding-left:10px;">Check that it is inside your <strong>wp-content</strong> directory and spelt correctly.</span></br>', 'my-text-domain' ); 		
+		echo _e( '<span style="font-size:14px; padding-left:10px;">The plugin was <strong>unable</strong> to create the <strong>prosperent-themes</strong> directory inside <strong>wp-content</strong>.</span><br><br>', 'my-text-domain' );
+		echo _e( '<span style="font-size:14px; padding-left:10px;">Please create a <strong>prosperent-themes</strong> directory inside <strong>wp-content</strong>.</span><br>', 'my-text-domain' );	
 		echo '</div>';		
+	}
+	else
+	{
+		wp_redirect( admin_url( 'admin.php?page=prosper_themes' ) );
+	}
 }
+
+
+echo __( '<ol><li>Place new theme directories inside <strong>prosperent-themes</strong>.<a href="#" class="prosper_tooltip"><img border="0" src="' . PROSPER_IMG . '/help.png"><span>New theme directories should contain one or all of the following: <br>&bull;&nbsp;<strong>product.php</strong><br>&bull;&nbsp;<strong>productPage.php</strong><br>&bull;&nbsp;a <strong>css file</strong> <br>(The CSS file can be named anything, the plugin is looking for the extension .css).</span></a></li><li>Select the desired theme.</li></ol>', 'prosperent-suite' );
+echo '<p class="prosper_desc">' . __( "", 'prosperent-suite' ) . '</p>';
+
 
 $prosperAdmin->adminFooter();
